+++ src/components/UserButton.tsx
@@ -3,23 +3,7 @@
 import Link from "next/link";
 import { useCookies } from "react-cookie";
 
-const jump = () => {
-	const clientId = process.env.DISCORD_CLIENT_ID;
-	const frontendAddress = process.env.NEXT_PUBLIC_FRONTEND_ADDRESS;
-
-	if (!clientId || !frontendAddress) {
-		console.error("環境変数が正しく設定されていません");
-		return;
-	}
-
-	window.open(
-		"https://discord.com/oauth2/authorize" +
-			`?client_id=${clientId}` +
-			"&response_type=code" +
-			`&redirect_uri=${frontendAddress}/login` +
-			"&scope=identify",
-	);
-};
+import { LoginButton } from "./LoginButton";
 
 export const UserButton = () => {
 	const [cookies] = useCookies<string>();
@@ -28,6 +12,6 @@ export const UserButton = () => {
 		const userId = "";
 		return <Link href={`/profile?u=${userId}`}>マイページ</Link>;
 	} else {
-		return <button onClick={jump}>ログイン</button>;
+		return <LoginButton label="Login" />;
 	}
 };
